libs/events: Move into internal
