+++ internal/consensus/state.go
@@ -16,8 +16,8 @@ import (
 	cfg "github.com/cometbft/cometbft/config"
 	"github.com/cometbft/cometbft/crypto"
 	cstypes "github.com/cometbft/cometbft/internal/consensus/types"
+	cmtevents "github.com/cometbft/cometbft/internal/events"
 	sm "github.com/cometbft/cometbft/internal/state"
-	cmtevents "github.com/cometbft/cometbft/libs/events"
 	"github.com/cometbft/cometbft/libs/fail"
 	cmtjson "github.com/cometbft/cometbft/libs/json"
 	"github.com/cometbft/cometbft/libs/log"
